Avoid integer conversion warning. A follow-up to r1872107.
* subversion/tests/libsvn_subr/mergeinfo-test.c
  (rangelist_to_array): Add explicit casts.

git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1872108 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/libsvn_subr/mergeinfo-test.c
@@ -1841,7 +1841,7 @@ rangelist_to_array(rl_array_t *a,
       svn_merge_range_t *range = APR_ARRAY_IDX(rl, i, svn_merge_range_t *);
       int r;
 
-      for (r = range->start + 1; r <= range->end; r++)
+      for (r = (int)range->start + 1; r <= (int)range->end; r++)
         {
           a->root[r] = TRUE;
           a->inherit[r] = range->inheritable;
